create the directory structure and increment Piece
I create the directory structure with just one file
because the git don't track empty folders.
The Piece and Pawn is incremental, but I don't know
what the IA team need as interface. Create main file.

please, if there's no objections, follow the comments
and the file structure(this is special to you, Wallace).

Questions : Maybe the position class need to be in utils
directory
time elapsed : 1h